requirements.txt
Pillow==8.0.1  
tkinterdnd2==1.0.3 
